// shared/bsdd-api/generated/types.gen.ts
export type ClassContractV1 = {
    /**
     * URI of the dictionary
     */
    dictionaryUri?: string | null;
    /**
     * Date of activation
     */
    activationDateUtc: string;
    /**
     * Code used originally by the dictionary
     */
    code: string;
    /**
     * Language code of the creator
     */
    creatorLanguageCode?: string | null;
    /**
     * List of countries where used
     */
    countriesOfUse?: Array<string> | null;
    /**
     * Country of origin
     */
    countryOfOrigin?: string | null;
    /**
     * Date of deactivation
     */
    deActivationDateUtc?: string | null;
    /**
     * Formal definition
     */
    definition?: string | null;
    /**
     * Explanation of the deprecation
     */
    deprecationExplanation?: string | null;
    /**
     * Description
     */
    description?: string | null;
    /**
     * Reference to a(n official) document
     */
    documentReference?: string | null;
    /**
     * Name
     */
    name: string;
    /**
     * Unique identification
     */
    uri: string;
    /**
     * List of codes of the replaced items
     */
    replacedObjectCodes?: Array<string> | null;
    /**
     * List of codes of the replacing items
     */
    replacingObjectCodes?: Array<string> | null;
    /**
     * Date of the revision
     */
    revisionDateUtc?: string | null;
    /**
     * Revision number
     */
    revisionNumber?: number | null;
    /**
     * Status, can be: Preview, Active or Inactive
     */
    status: string;
    /**
     * List of subdivisions (e.g. states) where used
     */
    subdivisionsOfUse?: Array<string> | null;
    /**
     * Alternative unique global identification
     */
    uid?: string | null;
    /**
     * Date of the version
     */
    versionDateUtc: string;
    /**
     * Version number
     */
    versionNumber?: number | null;
    /**
     * URI of a visual representation
     */
    visualRepresentationUri?: string | null;
    /**
     * Type of Class
     */
    classType?: string | null;
    /**
     * Code that can be used for dictionary specific purposes
     */
    referenceCode?: string | null;
    /**
     * List of synonyms for the class
     */
    synonyms?: Array<string> | null;
    /**
     * List of related IFC entity names (version independent)
     */
    relatedIfcEntityNames?: Array<string> | null;
    parentClassReference?: ClassReferenceContractV1;
    /**
     * List of the properties of this class
     */
    classProperties?: Array<ClassPropertyContractV1> | null;
    /**
     * List of relations to other classes, can be reference to classes of other dictionaries
     */
    classRelations?: Array<ClassRelationContractV1> | null;
    /**
     * List of child classes (only filled if requested)
     */
    childClassReferences?: Array<ClassReferenceContractV1> | null;
    /**
     * List of relations of other classes to this class (only filled if requested)
     */
    reverseClassRelations?: Array<ClassReverseRelationContractV1> | null;
    /**
     * List of parents of this class till root
     */
    hierarchy?: Array<HierarchyItemContractV1> | null;
};

export type ClassListItemContractV1Classes = {
    uri?: string | null;
    code?: string | null;
    name?: string | null;
    classType?: string | null;
    referenceCode?: string | null;
    parentClassCode?: string | null;
    /**
     * Description of the class, max 101 characters long
     */
    descriptionPart?: string | null;
    children?: Array<ClassListItemContractV1Classes> | null;
};
// ...
export type ClassReferenceContractV1 = {
    uri: string;
    name?: string | null;
    code?: string | null;
};

export type ClassRelationContractV1 = {
    uri?: string | null;
    /**
     * Can be one of: HasReference, IsEqualTo, IsSynonymOf, IsParentOf, IsChildOf, HasPart
     */
    relationType: string;
    /**
     * Namespace URI of the related class
     */
    relatedClassUri: string;
    /**
     * Name of the related class
     */
    relatedClassName?: string | null;
    /**
     * Optional provision of a fraction of the total amount (e.g. volume or weight) that applies to the class relations of one relation type
     */
    fraction?: number | null;
};
// ...
export type ClassReverseRelationContractV1 = {
    /**
     * Can be one of: HasReference, IsEqualTo, IsSynonymOf, IsParentOf, IsChildOf, HasPart
     */
    relationType: string;
    /**
     * URI of the reverse related class
     */
    classUri: string;
    /**
     * Name of the reverse related class
     */
    className?: string | null;
    /**
     * Optional provision of a fraction of the total amount (e.g. volume or weight) that applies to the class relations of one relation type
     */
    fraction?: number | null;
    /**
     * The URI of the dictionary that contains the reverse related class. This URI can be used to retrieve the dictionary
     */
    dictionaryUri?: string | null;
};
// ...
export type ClassSearchResultContractV1 = {
    /**
     * Name of the Class
     */
    name?: string | null;
    /**
     * Unique identification of the Class
     */
    uri?: string | null;
    /**
     * Code that can be used for domain specific purposes
     */
    referenceCode?: string | null;
    /**
     * Type of the class
     */
    classType?: string | null;
    definition?: string | null;
    synonyms?: Array<string> | null;
};
// ...
export type DictionaryClassesResponseContractV1Classes = {
    code: string;
    uri: string;
    name: string;
    version: string;
    organizationCodeOwner: string;
    organizationNameOwner: string;
    changeRequestEmail?: string | null;
    /**
     * The default language for this domain
     */
    defaultLanguageCode: string;
    isLatestVersion: boolean;
    isVerified: boolean;
    isPrivate: boolean;
    /**
     * Name or short description of the license under which you can use this data
     */
    license?: string | null;
    /**
     * URL where you can find more details about the license or default link to info about licensing
     */
    licenseUrl?: string | null;
    /**
     * Name or short description of the quality assurance procedure used while creating and maintaining the domain data
     */
    qualityAssuranceProcedure?: string | null;
    /**
     * URL where you can find more details about the quality assurance procedure
     */
    qualityAssuranceProcedureUrl?: string | null;
    /**
     * For state indication, like "Preview", "Active", "InActive", "Released
     */
    status?: string | null;
    /**
     * Url to site with more info on domain
     */
    moreInfoUrl?: string | null;
    /**
     * Date of the release of this version
     */
    releaseDate?: string | null;
    /**
     * Date and time the data of the domain has been last updated in bSDD
     */
    lastUpdatedUtc?: string;
    classes?: Array<ClassListItemContractV1Classes> | null;
    /**
     * Total number of classes within the dictionary.
     */
    classesTotalCount?: number | null;
    classesOffset?: number | null;
    /**
     * The returned number of classes.
     */
    classesCount?: number | null;
};
// ...
export type DictionarySearchResultContractV1 = {
    name?: string | null;
    uri?: string | null;
    classes?: Array<ClassSearchResultContractV1> | null;
};

export type HierarchyItemContractV1 = {
    level: number;
    name: string;
    code: string;
    uri: string;
};
// ...
export type SearchInDictionaryResponseContractV1 = {
    totalCount?: number;
    offset?: number;
    count?: number;
    dictionary?: DictionarySearchResultContractV1;
};